fix checkout
