+++ src/script/pages/Checkout/Components/CheckoutList.tsx
@@ -1,0 +1,17 @@
+
+
+
+export default function CheckoutList ({cart}) {
+    console.log(cart)
+    return (
+
+        <div className="Order">
+            <ul>
+            
+            </ul>
+            <div className="summery">
+                
+            </div>
+        </div>
+    )
+}
