+++ src/script/pages/Checkout/Components/CheckoutProduct.tsx
